Reimplement Column Frontend Wireframe
