+++ frontend/src/app/event/event-details/event-details.component.spec.ts
