+++ frontend/src/app/event/event-page/event-page.component.css
@@ -1,0 +1,60 @@
+.container {
+    padding-left: 24px;
+}
+
+.events-grid {
+    display: grid;
+    grid-template-columns: auto 560px;
+    /* grid-template-rows: 0fr 2fr; */
+    row-gap: 15px;
+    column-gap: 24px;
+    width: inherit;
+}
+
+.parameter-bar {
+    padding: 24px 24px 0px 24px;
+    display: flex;
+    direction: row;
+    align-items: flex-start;
+    justify-content: space-between;
+    z-index: 2;
+    position: sticky;
+    top: 65px;
+    background-color: #303030; /*fafafa*/
+}
+
+.search-bar {
+    width: 100%;
+    border-radius: 100%;
+}
+
+.mat-mdc-card {
+    margin: 0 !important;
+    max-width: 100%;
+}
+
+.details-column {
+    position: fixed;
+    right: 36px;
+    bottom: 24px;
+    top: 165px;
+    width: 548px;
+}
+
+@media only screen and (max-width: 1100px) {
+    .events-grid {
+        grid-template-columns: auto 400px;
+    }
+    .details-column {
+        width: 388px;
+    }
+}
+
+@media only screen and (max-width: 900px) {
+    .events-grid {
+        grid-template-columns: auto 0;
+    }
+    .details-column {
+        visibility: hidden;
+    }
+}
