+++ frontend/src/app/event/event-page/event-page.component.spec.ts
