+++ frontend/src/app/event/event-routing.module.ts
@@ -12,4 +12,4 @@ const routes: Routes = [
     imports: [RouterModule.forChild(routes)],
     exports: [RouterModule]
 })
-export class ExperimentalRoutingModule { }
+export class EventRoutingModule { }
