+++ frontend/src/app/event/widgets/event-card/event-card.widget.css
