+++ frontend/src/app/event/widgets/event-card/event-card.widget.html
@@ -1,0 +1,15 @@
+<mat-card appearance="outlined" style="margin-top: 12px !important;" routerLink="/events/detail">
+    <div style="display: flex; flex-direction: column; padding: 12px;">
+        <div style="display: flex; flex-direction: row;">
+            <div style="display: flex; flex-direction: column; width: 50%;">
+                <p style="font-weight: 500; margin: 0px;">{{eventName}}</p>
+                <p style="margin: 0px;">{{eventOrganizationName}}</p>
+            </div>
+            <div style="display: flex; flex-direction: column; width: 50%; justify-content: flex-end;">
+                <p style="margin: 0px; text-align: end;">{{eventTimeText}}</p>
+                <p style="margin: 0px; text-align: end;">{{eventLocation}}</p>
+            </div>
+        </div>
+        <p>Here is a short description about this event. This would have a character limit similar to the short description of an organization.</p>
+    </div>
+</mat-card>
