+++ frontend/src/app/event/widgets/event-detail-card/event-detail-card.widget.html
