--- frontend/src/app/experimental/event-page/event-page.component.css
@@ -1,33 +1,0 @@
-.container {
-    padding: 24px;
-}
-
-.events-grid {
-    display: grid;
-    /* grid-template-columns: 1fr 1fr; */
-    /* grid-template-rows: 0fr 2fr; */
-    row-gap: 15px;
-    column-gap: 24px;
-}
-
-.parameter-bar {
-    width: 100%;
-    display: flex;
-    direction: row;
-    align-items: flex-start;
-    justify-content: space-between;
-}
-
-.search-bar {
-    width: 100%;
-    border-radius: 100%;
-}
-
-.mat-mdc-card {
-    margin: 0 !important;
-    max-width: 100%;
-}
-
-.details-column {
-    height: 100%;
-}
