--- frontend/src/app/experimental/event-page/event-page.component.html
@@ -1,39 +1,0 @@
-<div class="container">
-
-    <div class="parameter-bar">
-        <search-bar class="search-bar" />
-    </div>
-
-    <div class="events-grid">
-
-        <div class="events-column">
-            <mat-card class="events-section" appearance="outlined" style="padding: 16px;">
-                <mat-card-title>October 28th</mat-card-title>
-                <mat-divider style="padding-bottom: 8px;" />
-
-                <event-card id="firstcard" eventName="HackNC Hackathon" eventOrganizationName="HackNC"
-                    eventTimeText="10:00 AM - All day" eventLocation="Great Hall (Student Union)" />
-
-                <event-card eventName=" Introduction to APIs Workshop" eventOrganizationName="CADS"
-                    eventTimeText="7:00 PM - 1 hour" eventLocation="Sitterson Hall - SN014" />
-
-                <event-card eventName="Front End Design with Tailwind CSS" eventOrganizationName="CSXL"
-                    eventTimeText="8:00 PM - 1 hour" eventLocation="Sitterson Hall - SN155" />
-
-
-                <mat-card-title style="padding-top: 12px;">October 29th</mat-card-title>
-                <mat-divider style="padding-bottom: 8px;" />
-
-                <event-card eventName="HackNC Hackathon" eventOrganizationName="HackNC"
-                    eventTimeText="10:00 AM - All Day" eventLocation="Great Hall (Student Union)" />
-
-                <mat-card-title style="padding-top: 12px;">October 30th</mat-card-title>
-                <mat-divider style="padding-bottom: 8px;" />
-
-                <event-card eventName="Fall Career Fair" eventOrganizationName="Department of Computer Science"
-                    eventTimeText="12:00 PM - 3 hours" eventLocation="Sitterson Lower Lobby" />
-
-            </mat-card>
-        </div>
-    </div>
-</div>
