--- frontend/src/app/experimental/event-page/event-page.component.ts
@@ -1,30 +1,0 @@
-import { Component } from '@angular/core';
-import { Organization } from 'src/app/organization/organization.service';
-import { profileResolver } from 'src/app/profile/profile.resolver';
-
-@Component({
-  selector: 'app-event-page',
-  templateUrl: './event-page.component.html',
-  styleUrls: ['./event-page.component.css']
-})
-export class EventPageComponent {
-
-  /** Route information to be used in App Routing Module */
-  public static Route = {
-    path: '',
-    title: 'Events',
-    component: EventPageComponent,
-    canActivate: [],
-    resolve: { profile: profileResolver }
-  }
-}
-
-export interface EventDetailData {
-  name: string;
-  organization: Organization;
-  startText: string;
-  endText: string;
-  location: string;
-  description: string;
-  requiresPreregistration: boolean;
-}
