--- frontend/src/app/experimental/widgets/event-card/event-card.widget.html
@@ -1,12 +1,0 @@
-<mat-card appearance="outlined" style="margin-top: 12px !important;" routerLink="/events/detail">
-    <div style="display: flex; flex-direction: row; padding: 12px;">
-        <div style="display: flex; flex-direction: column; width: 50%;">
-            <p style="font-weight: 500; margin: 0px;">{{eventName}}</p>
-            <p style="margin: 0px;">{{eventOrganizationName}}</p>
-        </div>
-        <div style="display: flex; flex-direction: column; width: 50%; justify-content: flex-end;">
-            <p style="margin: 0px; text-align: end;">{{eventTimeText}}</p>
-            <p style="margin: 0px; text-align: end;">{{eventLocation}}</p>
-        </div>
-    </div>
-</mat-card>
